En proceso carta bootcamp 25, carta se destruye una vez usada (correcto)
