+++ Videojuego/01_Prototipo_BitDestroyer/Assets/Scripts/BootCampScripts/BootcampCard_25.cs
@@ -14,103 +14,31 @@ using UnityEngine;
 public class BootcampCard_25 : MonoBehaviour
 {
     /* Panel to show the roulette, initially active false in the spawn_card_in_hand script */
-    private GameObject roulettePanel;
+    public GameObject roulettePanel;
     public bool activePanelRoulette = false; 
-    public float initialRotationSpeed = 1000f; // Velocidad de rotación inicial
-    private float currentRotationSpeed; // Velocidad de rotación actual
-    private bool isSpinning = false;    // State of the roulette
-    private float spinDuration = 3.0f;  // Duración del giro
-
-    private GameObject cardObject; // Objeto de la tarjeta
-
+    
     void Start()
     {
-        Debug.Log("BootcampCard_25.cs STARTED!");
-        roulettePanel = GameObject.Find("RoulettePanel");
+        Debug.Log("BootcampCard_25.cs STARTED IN SLOT!");
 
-        // Listar todos los objetos a los que se ha asignado este script
-        BootcampCard_25[] allInstances = FindObjectsOfType<BootcampCard_25>();
-        Debug.Log("Total objects with BootcampCard_25 script: " + allInstances.Length);
-
-        foreach (var instance in allInstances)
-        {
-            Debug.Log("Object name: " + instance.gameObject.name + ", Tag: " + instance.gameObject.tag);
-
-            // Guardar el objeto con la etiqueta "Card"
-            if (instance.gameObject.CompareTag("Card"))
-            {
-                cardObject = instance.gameObject;
-                Debug.Log("Card object found: " + cardObject.name);
-            }
+        if (roulettePanel == null) {
+            roulettePanel = GameObject.Find("RoulettePanel");
+        } 
+        else {
+            Debug.Log("Roulette Panel already assigned!");
         }
     }
     void Update()
     {
-        // if (activePanelRoulette && roulettePanel != null) {
-        //      roulettePanel.SetActive(true);
-        // } 
-        // else {
-        //     Debug.Log("Roulette Panel is null");
-        // }
-
-        if (isSpinning)
+        if (activePanelRoulette)
         {
-            this.transform.Rotate(Vector3.forward, -currentRotationSpeed * Time.deltaTime);
+            roulettePanel.SetActive(true);
+            ToDestroy();
         }
     }
 
-    public void StartSpin()
+    private void ToDestroy()
     {
-        if (!isSpinning)
-        {
-            isSpinning = true;
-            currentRotationSpeed = initialRotationSpeed;
-            StartCoroutine(StopSpin());
-        }
-    }
-
-    private IEnumerator StopSpin()
-    {
-        float elapsedTime = 0f;
-
-        while (elapsedTime < spinDuration)
-        {
-            elapsedTime += Time.deltaTime;
-            currentRotationSpeed = Mathf.Lerp(initialRotationSpeed, 0, elapsedTime / spinDuration);
-            yield return null;
-        }
-
-        isSpinning = false;
-        DetermineResult();
-    }
-
-    private void DetermineResult()
-    {
-        float zRotation = transform.eulerAngles.z % 360;
-        string result;
-
-        if (zRotation < 90)
-            result = "Rojo";
-        else if (zRotation < 180)
-            result = "Azul";
-        else if (zRotation < 270)
-            result = "Rojo";
-        else
-            result = "Azul";
-
-        Debug.Log("Resultado: " + result);
-
-        roulettePanel.SetActive(false);
-
-        if (cardObject != null)
-        {
-            Debug.Log("Destroying card object: " + cardObject.name);
-            Destroy(cardObject);
-            Debug.Log("BootcampCard_25.cs DESTROYED!");
-        }
-        else
-        {
-            Debug.Log("Card object not found.");
-        }
+        Destroy(gameObject);
     }
 }
