+++ Videojuego/01_Prototipo_BitDestroyer/Assets/Scripts/BootCampScripts/RouletteManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1e25bcca68d8c41bbb9f157d4a2d19f9
+guid: de5ffb957ba3d480c9d6aacdf7104115
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
