+++ Videojuego/01_Prototipo_BitDestroyer/Assets/Scripts/spawn_card_in_hand.cs
@@ -55,7 +55,7 @@ public class CardSpawner : MonoBehaviour
     {
         /* Hide the panel at tne beggining of scene */
         panelElectionCards.SetActive(false);
-        //panelRoulette.SetActive(false);
+        panelRoulette.SetActive(false);
 
         /* Add listener to the buttons to select one card */
         // cardOneElection.onClick.AddListener(() => {
@@ -255,6 +255,7 @@ public class CardSpawner : MonoBehaviour
                 switch (cardData.id) {
                     case 25:
                         var bootcampScript = newCard.AddComponent<BootcampCard_25>();
+                        bootcampScript.roulettePanel = panelRoulette;
                         break;
                 }
                 break;
